Make text decoding an optional feature in rten-generate
This allows downstream projects to use a different tokenization library (eg.
tiktoken, HuggingFace tokenizers) and avoid pulling in rten-text and its
dependencies (regex engine etc.).

In the process `TextGenerator` was renamed to `TextDecoder` to align with
`GeneratorUtils::decode`.
