+++ rten-examples/Cargo.toml
@@ -17,7 +17,7 @@ png = "0.17.6"
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 rten = { path = "../", features = ["mmap", "random"] }
-rten-generate = { path = "../rten-generate" }
+rten-generate = { path = "../rten-generate", features=["text-decoder"] }
 rten-imageio = { path = "../rten-imageio" }
 rten-imageproc = { path = "../rten-imageproc" }
 rten-tensor = { path = "../rten-tensor" }
